+++ sdk/src/main/java/ly/count/android/sdk/ModuleCrash.java
@@ -33,6 +33,8 @@ public class ModuleCrash extends ModuleBase {
     @Nullable
     Map<String, String> metricOverride = null;
 
+    BreadcrumbHelper breadcrumbHelper;
+
     ModuleCrash(Countly cly, CountlyConfig config) {
         super(cly, config);
         L.v("[ModuleCrash] Initialising");
@@ -47,6 +49,9 @@ public class ModuleCrash extends ModuleBase {
         metricOverride = config.metricOverride;
 
         crashesInterface = new Crashes();
+        breadcrumbHelper = new BreadcrumbHelper(config.sdkInternalLimits.maxBreadcrumbCount, L);
+
+        assert breadcrumbHelper != null;
     }
 
     /**
@@ -117,7 +122,7 @@ public class ModuleCrash extends ModuleBase {
         sendCrashReportToQueue(dumpString, false, true, null);
     }
 
-    public void sendCrashReportToQueue(String error, boolean nonfatal, boolean isNativeCrash, @Nullable final Map<String, Object> customSegmentation) {
+    public void sendCrashReportToQueue(@NonNull String error, boolean nonfatal, boolean isNativeCrash, @Nullable final Map<String, Object> customSegmentation) {
         L.d("[ModuleCrash] sendCrashReportToQueue");
 
         Map<String, Object> combinedSegmentationValues = new HashMap<>();
@@ -271,12 +276,8 @@ public class ModuleCrash extends ModuleBase {
      * @param itIsHandled If the exception is handled or not (fatal)
      * @return Returns link to Countly for call chaining
      */
-    Countly recordExceptionInternal(final Throwable exception, final boolean itIsHandled, final Map<String, Object> customSegmentation) {
+    Countly recordExceptionInternal(@Nullable final Throwable exception, final boolean itIsHandled, final Map<String, Object> customSegmentation) {
         L.i("[ModuleCrash] Logging exception, handled:[" + itIsHandled + "]");
-
-        if (!_cly.isInitialized()) {
-            throw new IllegalStateException("Countly.sharedInstance().init must be called before recording exceptions");
-        }
 
         if (!consentProvider.getConsent(Countly.CountlyFeatureNames.crashes)) {
             return _cly;
@@ -284,7 +285,6 @@ public class ModuleCrash extends ModuleBase {
 
         if (exception == null) {
             L.d("[ModuleCrash] recordException, provided exception was null, returning");
-
             return _cly;
         }
 
@@ -321,11 +321,11 @@ public class ModuleCrash extends ModuleBase {
         }
 
         if (breadcrumb == null || breadcrumb.isEmpty()) {
-            L.e("[Crashes] Can't add a null or empty crash breadcrumb");
+            L.w("[ModuleCrash] addBreadcrumbInternal, Can't add a null or empty crash breadcrumb");
             return _cly;
         }
 
-        DeviceInfo.addLog(breadcrumb, _cly.config_.sdkInternalLimits.maxBreadcrumbCount, _cly.config_.sdkInternalLimits.maxValueSize);
+        breadcrumbHelper.addBreadcrumb(breadcrumb, _cly.config_.sdkInternalLimits.maxValueSize);
         return _cly;
     }