+++ sdk/src/main/java/ly/count/android/sdk/ModuleRequestQueue.java
@@ -7,7 +7,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.json.JSONObject;
 
 public class ModuleRequestQueue extends ModuleBase implements BaseInfoProvider {
     RequestQueue requestQueueInterface;
@@ -71,8 +70,15 @@ public class ModuleRequestQueue extends ModuleBase implements BaseInfoProvider {
         try {
             List<String> filteredRequests = new ArrayList<>();
 
-            if (storedRequests == null || targetAppKey == null) {
+            if (storedRequests == null) {
                 //early abort
+                L.w("[ModuleRequestQueue] requestQueueReplaceWithAppKey, stopping replacing due to stored requests being 'null'");
+                return filteredRequests;
+            }
+
+            if (targetAppKey == null || targetAppKey.isEmpty()) {
+                //early abort
+                L.w("[ModuleRequestQueue] requestQueueReplaceWithAppKey, stopping replacing due to target app key being 'null' or empty string");
                 return filteredRequests;
             }
 
@@ -148,7 +154,7 @@ public class ModuleRequestQueue extends ModuleBase implements BaseInfoProvider {
         int eventsInEventQueue = storageProvider.getEventQueueSize();
         L.v("[ModuleRequestQueue] forceSendingEvents, forced:[" + forceSendingEvents + "], event count:[" + eventsInEventQueue + "]");
 
-        if ((forceSendingEvents && eventsInEventQueue > 0) || eventsInEventQueue >= Countly.EVENT_QUEUE_SIZE_THRESHOLD) {
+        if ((forceSendingEvents && eventsInEventQueue > 0) || eventsInEventQueue >= _cly.EVENT_QUEUE_SIZE_THRESHOLD) {
             requestQueueProvider.recordEvents(storageProvider.getEventsForRequestAndEmptyEventQueue());
         }
     }
@@ -227,6 +233,11 @@ public class ModuleRequestQueue extends ModuleBase implements BaseInfoProvider {
      * Send request data after removing the predefined keys
      */
     synchronized public void addDirectRequestInternal(@NonNull Map<String, String> requestMap) {
+        Long pccTsStartAddDirectRequest = 0L;
+        if (pcc != null) {
+            pccTsStartAddDirectRequest = UtilsTime.getNanoTime();
+        }
+
         L.i("[ModuleRequestQueue] Calling addDirectRequestInternal");
         if (!_cly.isInitialized()) {
             L.e("Countly.sharedInstance().init must be called before adding direct request, returning");
@@ -276,6 +287,10 @@ public class ModuleRequestQueue extends ModuleBase implements BaseInfoProvider {
             L.w("[ModuleRequestQueue] addDirectRequest, [" + delta + "] restricted keys are removed");
         }
         requestQueueProvider.sendDirectRequest(filteredRequestMap);
+
+        if (pcc != null) {
+            pcc.TrackCounterTimeNs("ModuleRequestQueue_addDirectRequestInternal", UtilsTime.getNanoTime() - pccTsStartAddDirectRequest);
+        }
     }
 
     void esWriteCachesToPersistenceInternal(@Nullable ExplicitStorageCallback callback) {